feat(projects): Add GemCommit project to portfolio
This commit adds the GemCommit project, a VS Code extension that generates commit messages automatically using Google Gemini AI. The project includes a description, technology used (TypeScript), and a link to the Visual Studio Code Marketplace.  It also includes a thumbnail for the project and expand the card to fill the full width on large screens.